Refactor game architecture by removing deprecated day/night cycle components and transitioning to a wave system management. Update relevant game state properties and serialization methods to reflect the new structure. Introduce new documentation for extension field-level delta compression analysis.

+++ packages/game-client/src/extensions/client-entity.ts
@@ -210,7 +210,7 @@ export abstract class ClientEntityBase {
       console.warn(`Entity type mismatch: expected ${this.type}, got ${type}`);
     }
 
-    const fieldCount = currentReader.readUInt32();
+    const fieldCount = currentReader.readUInt8();
     for (let i = 0; i < fieldCount; i++) {
       const fieldName = currentReader.readString();
       const valueType = currentReader.readUInt32();
@@ -239,7 +239,7 @@ export abstract class ClientEntityBase {
       (this as any)[fieldName] = value;
     }
 
-    const extensionCount = currentReader.readUInt32();
+    const extensionCount = currentReader.readUInt8();
     const existingExtensions = new Map<string, ClientExtension>();
     for (const ext of this.extensions) {
       const extType = (ext.constructor as any).type;
